Many updates and bug fixes in the Diabetes model
- Added DeathSimpleCHDSubmodel to add death to CHD acute events
- Added EmpiricalSpainDeathSubmodel to use an empirical instead of a
Gompertz function for all-causes death. This improves accuracy and
performance
- Added LySecondOrderParams to mimic Conget's model from Core
- Changed classpath to make the HTA classes depend from the project and
not a jar file.
- Added "dead" attribute and the corresponding getter and setters to the
Patient class to represent whether the patient is dead.
- Removed T1DMPatientInfo.Type.FINISH since it was not working as
expected. It was only called only immediately after
T1DMPatientInfo.Type.DEATH, but never if the simulation stopped before
the patient's death.
- Fixed inforeceivers to 1) use SimulationTimeInfo instead of
SimulationEndInfo; 2) Do not use the FINISH type; 3) Fill the
information when the simulation ends and the patient has not died
- Added SIMLENGTH, USE_CHD_DEATH_MODEL, and INIT_PROP to
BasicConfigParams
- Changed T1DMSimulation to use BasicConfigParams.SIMLENGTH as the
length of the simulation
- Added initial proportion of BGRET to SheffieldRETSubmodel
- Added initial proportion of ALB1 to SheffieldNPHSubmodel
- Added a method to get the generated patients from any HTASimulation
- Removed T1DMPatient.FinalizeEvent. Now simply uses
DiscreteEvent.DefaultFinalizeEvent
- Added the possibility to cause death to chronic complications. Fixed
T1DMPatient, T1DMProgressionPair, and T1DMProgression in consequence
- Added the cause of the death to the T1DMPatient.DeathEvent
- Modified UnconsciousSecondOrderParams to use DeathSimpleCHDSubmodel
- Modified T1DMMain to: 1) allow printing individual outcomes; 2) use
just a "population" option to select the set of second order params; 3)
allow dynamically selecting initial proportions for health states; 4)
allow indicating a simulation length; 5) allow selecting the CHD
submodel with deaths