Several changes to make loading scenarios from ontologies easier
The most important change involves that registering a disease,
manifestation, intervention, population or death submodel in a
repository is now separated from registering their second order
parameters. The former is performed once creating an instance (disease,
manifestation, intervention) or using the specific setter in the
repository (population, death submodel); whereas the latter is performed
by using a new method (registerAllSecondOrderParams) that should be
invoked  after all the components are created.
The method has been tested with the hardcoded diabetes model and the
ontology-based PBD model