Some refactoring and included first type of modifications
- Moved auxiliary classes to package "utils"
- Included modifications to the probability of a pathway due to an
intervention
- Now the base utility is defined in the Population and then loaded into
the Repository. It is supposed to be a single parameter, and its default
value is 1
- The UtilityCalculators now use the base utility in the repository as
the reference utility, instead of passing a constant parameter (thus,
such utility can be used now as a second order parameter)
- Diseases now allow to have utilities apart from disutilities (the
method SecondOrderParamsRepository.getDisutilityForDisease() was not
working properly before)
- Added a map<String, Manifestation> to help searching Manifestations by
its name. Added a method to do so.
- Now Developments are also registered in the Repository
- Complete refactoring of data properties, object properties and classes
to represent modifications in the ontology