Improved functionality of OwlHelper and full refactoring of ontology
- Now all the classes that define elements in a model are subclasses of
ModelItem
- Now Characteristics are Attributes
- Modifications are not anymore classes but is an object property of
Valuable (attribute values and parameter) instances
- Included MeasuredDataItemType and Currency as DataItemType subclasses,
with data property hasUnit
- Included Model class to be able to define model projects and the
subset of instances that are included in such project (by means of the
usedByModel object property)
- Included subclasses of the Model class for decision tree, DES... The
idea is to be able, in the future, of automatizally selecting the best
type of model
- The knowledge on GuidelineRange is now integrated into Guideline, and
Guidelines can use recursive object properties to represent complex
behaviours.
- The knowledge on ScreeningTest and DiagnosisTest is now integrated
into the corresponding strategies.
- Almost all data properties that referred to knowledge susceptible of
having related uncertainty are now object properties with range Valuable
- hasLifeExpectancy is now an object property with domain Disease,
Development or Population
- hasMortalityFactor knowledge has been split into two object
properties: one with the same name and a "hasLifeTimeReduction" one
- Uncertainty is included by means of the hasParameterUncertainty,
hasStochasticUncertainty and hasHeterogeneity object properties, with
range Valuable
- The value of a Valuable instance is defined in the hasExpression data
property
- ManifestationPathway is now only useful to define conditional
pathways, e.g. pathway related to a previous Manifestation. Otherwise,
the pathway can be defined straightforward in the Manifestation.
- The pathway itself is defined with the "hasRiskCharacterization"
object property, with range Parameter
- Most object properties are now related to their domain classes by
means of a subclass relationship, as it has been done to data properties
- Some other minor fixes