Changed all the methods related to SecondOrderParam to use Patient
Before, most of these methods received an integer (generally, the
simulation id) or the simulation, but it was inconsistent, and leave no
place for increasing the flexibility of such methods (for example,
combine first and second order uncertainty).
The method SecondOrderParam.getValues(int) has been left to allow
SecondOrderParamsRepository.print(int) work properly as the whole
mechanism to read parameters is improved.